plausibly working refactor
